Move PauseDuration to domain/model/pause package
Move PauseDuration class from com.scalar.admin.kubernetes to
com.scalar.admin.kubernetes.domain.model.pause package to establish
proper domain layer structure.

Changes:
- Move PauseDuration.java using git mv (preserves git history)
- Update package declaration
- Make constructor public (required for cross-package access)
- Update imports in Pauser, Cli, Result, and PauserTest

No functional changes. Implementation remains unchanged.
